+++ CloudHarvestCorePluginManager/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module exposes key functionality for the CloudHarvestCorePluginManager package.
+"""
+
+from .decorators import register_definition
+from .functions import register_all
+from .registry import Registry